Removed circular dependency

+++ src/main/java/com/synopsys/integration/alert/common/descriptor/config/context/ChannelDistributionDescriptorActionApi.java
@@ -27,6 +27,7 @@ import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 
@@ -39,7 +40,6 @@ import com.synopsys.integration.alert.channel.event.DistributionEvent;
 import com.synopsys.integration.alert.common.ContentConverter;
 import com.synopsys.integration.alert.common.configuration.FieldAccessor;
 import com.synopsys.integration.alert.common.database.FieldConfigurationAccessor;
-import com.synopsys.integration.alert.common.descriptor.DescriptorMap;
 import com.synopsys.integration.alert.common.descriptor.ProviderDescriptor;
 import com.synopsys.integration.alert.common.descriptor.config.ui.CommonDistributionUIConfig;
 import com.synopsys.integration.alert.common.descriptor.config.ui.ProviderDistributionUIConfig;
@@ -58,13 +58,14 @@ public abstract class ChannelDistributionDescriptorActionApi extends DescriptorA
     private final FieldConfigurationAccessor configurationAccessor;
     private final Logger logger = LoggerFactory.getLogger(getClass());
     private final ContentConverter contentConverter;
-    private final DescriptorMap descriptorMap;
+    private final List<ProviderDescriptor> providerDescriptors;
 
-    public ChannelDistributionDescriptorActionApi(final DistributionChannel distributionChannel, final FieldConfigurationAccessor configurationAccessor, final ContentConverter contentConverter, final DescriptorMap descriptorMap) {
+    public ChannelDistributionDescriptorActionApi(final DistributionChannel distributionChannel, final FieldConfigurationAccessor configurationAccessor, final ContentConverter contentConverter,
+        final List<ProviderDescriptor> providerDescriptors) {
         this.distributionChannel = distributionChannel;
         this.configurationAccessor = configurationAccessor;
         this.contentConverter = contentConverter;
-        this.descriptorMap = descriptorMap;
+        this.providerDescriptors = providerDescriptors;
     }
 
     @Override
@@ -151,22 +152,32 @@ public abstract class ChannelDistributionDescriptorActionApi extends DescriptorA
 
     @Override
     public void saveConfig(final FieldModel fieldModel) {
-        DescriptorActionApi providerActionApi = getProviderActionApi(fieldModel);
-        providerActionApi.saveConfig(fieldModel);
+        final DescriptorActionApi providerActionApi = getProviderActionApi(fieldModel);
+        if (null != providerActionApi) {
+            providerActionApi.saveConfig(fieldModel);
+        }
         super.saveConfig(fieldModel);
     }
 
     @Override
     public void deleteConfig(final FieldModel fieldModel) {
         final DescriptorActionApi descriptorActionApi = getProviderActionApi(fieldModel);
-        descriptorActionApi.deleteConfig(fieldModel);
+        if (null != descriptorActionApi) {
+            descriptorActionApi.deleteConfig(fieldModel);
+        }
         super.deleteConfig(fieldModel);
     }
 
-    private DescriptorActionApi getProviderActionApi(FieldModel fieldModel) {
+    private DescriptorActionApi getProviderActionApi(final FieldModel fieldModel) {
         final FieldAccessor fieldAccessor = fieldModel.convertToFieldAccessor();
         final String providerName = fieldAccessor.getString(CommonDistributionUIConfig.KEY_PROVIDER_NAME);
-        ProviderDescriptor providerDescriptor = descriptorMap.getProviderDescriptor(providerName);
-        return providerDescriptor.getRestApi(ConfigContextEnum.DISTRIBUTION);
+        final Optional<ProviderDescriptor> foundProviderDescriptor = providerDescriptors.stream()
+                                                                         .filter(providerDescriptor -> providerDescriptor.getName().equals(providerName))
+                                                                         .findFirst();
+        if (foundProviderDescriptor.isPresent()) {
+            return foundProviderDescriptor.get().getRestApi(ConfigContextEnum.DISTRIBUTION);
+        }
+
+        return null;
     }
 }
